Second Commit

+++ power.java
@@ -8,7 +8,7 @@ class power{
 	 if (n>=0 && n<31){
 		 for(int i=0;i<n;i++)
 		 {
-				System.out.println(Math.pow(2,i));
+				System.out.println((long)Math.pow(2,i));
 		}
 		 
 		 }
